requirements.txt
Django~=4.2.3
djangorestframework~=3.14.0
pytest~=7.4.0